Refactor connection handling and token generation logic
Moved connection handling functions to a new dedicated `conn.go` file for better modularization and clarity. Simplified token TTL handling by replacing redundant logic with a single default and updated schema types. This improves code maintainability and reduces duplication.
