+++ pkg/api/schema.go
@@ -2,11 +2,11 @@ package api
 
 type GenerateTokenRequest struct {
 	KeyID string `json:"key_id"`
-	TTL   int64  `json:"ttl"`
+	TTL   int    `json:"ttl"`
 }
 
 type GenerateTokenResponse struct {
 	Token string `json:"token"`
 	KeyID string `json:"key_id"`
-	TTL   int64  `json:"ttl"`
+	TTL   int    `json:"ttl"`
 }
